@@ -1,77 +1,47 @@
-import { Component } from 'react';
-import MarvelService from '../../services/MarvelService'
+import { useState, useEffect } from 'react';
+import useMarvelService from '../../services/MarvelService'
 import ErrorMessage from '../errorMessage/ErrorMessage';
 import Spinner from '../spinner/Spinner'
 import Skeleton from '../skeleton/Skeleton'
 import PropTypes from 'prop-types'; 
 import './charInfo.scss';
 
-class CharInfo extends Component {
-    state = {
-        char: null,
-        loading: false,
-        error: false
+const CharInfo = (props) => {
+
+    const [char, setChar] = useState(null)
+
+    const {loading, error, getCharacter, clearError} = useMarvelService();
+
+    useEffect(() => {
+        updateChar()
+    },[props.charId])
+
+    const onCharLoaded = (char) => {
+        setChar(char)
     }
 
-    marvelService = new MarvelService();
-
-    componentDidMount() {
-        this.updateChar()
-    }
-
-    componentDidUpdate(prevProsp) {
-        if (this.props.charId !== prevProsp.charId){
-            this.updateChar()
-        }
-    }
-
-    onCharLoaded = (char) => {
-        this.setState({
-            char, 
-            loading: false
-        })
-    }
-
-    onCharLoading = () => {
-        this.setState({
-            loading: true           
-        })
-    }
-
-    onError = () => {
-        this.setState({
-            loading: false, 
-            error: true
-        })
-    }
-
-    updateChar = () => {
-        const {charId} = this.props
+    const updateChar = () => {
+        const {charId} = props
         if (!charId) return
 
-        this.onCharLoading()
-        this.marvelService
-            .getCharacter(charId)
-            .then(this.onCharLoaded)
-            .catch(this.onError)
+        clearError()
+        getCharacter(charId)
+            .then(onCharLoaded)
     }
+    
+    const skeleton = char || loading || error ? null : <Skeleton/> 
+    const errorMessage = error ? <ErrorMessage/> : null
+    const spinner = loading ? <Spinner/> : null
+    const content = !(loading || error || !char) ? <View char={char}/> : null
 
-    render(){
-        const {char, loading, error} = this.state
-        const skeleton = char || loading || error ? null : <Skeleton/> 
-        const errorMessage = error ? <ErrorMessage/> : null
-        const spinner = loading ? <Spinner/> : null
-        const content = !(loading || error || !char) ? <View char={char}/> : null
-
-        return (
-            <div className="char__info">
-                {skeleton}
-                {errorMessage}
-                {spinner}
-                {content}
-            </div>
-        )
-    }
+    return (
+        <div className="char__info">
+            {skeleton}
+            {errorMessage}
+            {spinner}
+            {content}
+        </div>
+    )    
 }
 
 const View = ({char}) => {
@@ -117,7 +87,7 @@ const View = ({char}) => {
     )
 }
 
-CharInfo.prosType = {
+CharInfo.propTypes = {
     charId: PropTypes.number
 }
 
