@@ -1,94 +1,63 @@
-import { Component } from 'react';
-import MarvelService from '../../services/MarvelService'
+import { useState, useEffect } from 'react';
+import useMarvelService from '../../services/MarvelService'
 import ErrorMessage from '../errorMessage/ErrorMessage';
 import Spinner from '../spinner/Spinner'
+import setContent from '../../utils/setContent';
 import './randomChar.scss';
 import mjolnir from '../../resources/img/mjolnir.png';
 
-class RandomChar extends Component {
-    // constructor(props){
-    //     super(props)    
-    //     this.updateChar()
-    // }
+const RandomChar = () => {
+
+    const [char, setChar] = useState({})
+    const {loading, error, getCharacter, clearError, process, setProcess} = useMarvelService();
     
-    state = {
-        char: {},
-        loading: true,
-        error: false
+    useEffect(() => {
+        updateChar()
+    },[])
+
+    const onCharLoaded = (char) => {
+        setChar(char)
     }
 
-    marvelService = new MarvelService();
-
-    componentDidMount(){
-        this.updateChar()
-        // this.timerId = setInterval(this.updateChar, 10000)
-    }
-
-    componentWillUnmount() {
-        // clearInterval(this.timerId)
-    }
-
-    onCharLoaded = (char) => {
-        this.setState({
-            char, 
-            loading: false
-        })
-    }
-
-    onCharLoading = () => {
-        this.setState({
-            loading: true           
-        })
-    }
-
-    onError = () => {
-        this.setState({
-            loading: false, 
-            error: true
-        })
-    }
-
-    updateChar = () => {
+    const updateChar = () => {
+        clearError()
         const id = Math.floor(Math.random() * (1011400 - 1011000) + 1011000)
-        this.onCharLoading()
-        this.marvelService
-            .getCharacter(id)
-            .then(this.onCharLoaded)
-            .catch(this.onError)
+        getCharacter(id)
+            .then(onCharLoaded)
+            .then(setProcess('success'))
     }
 
-    render() {
-        const {char, loading, error} = this.state
-        const errorMessage = error ? <ErrorMessage/> : null
-        const spinner = loading ? <Spinner/> : null
-        const content = !(loading || error) ? <View char={char}/> : null
-     
-        return (
-            <div className="randomchar">
-               {errorMessage}
-               {spinner}
-               {content}
-                <div className="randomchar__static">
-                    <p className="randomchar__title">
-                        Random character for today!<br/>
-                        Do you want to get to know him better?
-                    </p>
-                    <p className="randomchar__title">
-                        Or choose another one
-                    </p>
-                    <button className="button button__main">
-                        <div className="inner" onClick={this.updateChar}>try it</div>
-                    </button>
-                    <img src={mjolnir} alt="mjolnir" className="randomchar__decoration"/>
-                </div>
+    // const errorMessage = error ? <ErrorMessage/> : null
+    // const spinner = loading ? <Spinner/> : null
+    // const content = !(loading || error) ? <View char={char}/> : null
+    
+    return (
+        <div className="randomchar">
+            {setContent(process, View, char)}
+            {/* {errorMessage}
+            {spinner}
+            {content} */}
+            <div className="randomchar__static">
+                <p className="randomchar__title">
+                    Random character for today!<br/>
+                    Do you want to get to know him better?
+                </p>
+                <p className="randomchar__title">
+                    Or choose another one
+                </p>
+                <button className="button button__main">
+                    <div className="inner" onClick={updateChar}>try it</div>
+                </button>
+                <img src={mjolnir} alt="mjolnir" className="randomchar__decoration"/>
             </div>
-        )
-    }
+        </div>
+    )
+
 }
 
-const View = ({char}) => {
-    const {name, description, thumbnail, homepage, wiki} = char
-    const fit = thumbnail.includes('image_not_available') ? 'randomchar__img__not' : 'randomchar__img'
+const View = ({data}) => {
+    const {name, description, thumbnail, homepage, wiki} = data
+    const fit = thumbnail === ('http://i.annihil.us/u/prod/marvel/i/mg/b/40/image_not_available.jpg') ? 'randomchar__img__not' : 'randomchar__img'
     
     return (
         <div className="randomchar__block">
