@@ -1,19 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import App from './components/app/App';
 import './style/style.scss';
 
-    // import MarvelService from './services/MarvelService';
 
-    // const marvelService = new MarvelService();
-
-    // marvelService.getAllCharacters().then(res => res.data.results.forEach(item => console.log(item.name)))
-    // marvelService.getCharacter(1011052).then(res => console.log(res))
-
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+ReactDOM
+  .createRoot(document.getElementById('root'))
+  .render(
+    // <React.StrictMode>
+      <App />
+    // </React.StrictMode>,    
+  );
 
