@@ -1,141 +1,109 @@
-import { Component } from 'react';
-import MarvelService from '../../services/MarvelService'
+import { useState, useEffect, useRef } from 'react';
+import { CSSTransition, TransitionGroup } from 'react-transition-group';
+import useMarvelService from '../../services/MarvelService'
 import ErrorMessage from '../errorMessage/ErrorMessage';
 import Spinner from '../spinner/Spinner'
 import PropTypes from 'prop-types'; 
 import './charList.scss';
 
-class CharList extends Component {
+const CharList = (props) => {
 
-    state = {
-        charList: [],
-        loading: true,
-        error: false,
-        newItemLoading: false,
-        offset: 210,
-        charEnded: false,
-    }
+    const [charList, setCharList] = useState([])
+    const [newItemLoading, setNewItemLoading] = useState(false)
+    const [offset, setOffset] = useState(210)
+    const [charEnded, setCharEnded] = useState(false)
+    
+    const {loading, error, getAllCharacters} = useMarvelService();
 
-    marvelService = new MarvelService();
+    useEffect(() => {
+        updateCharList(offset, true)
+    },[])
 
-    componentDidMount(){
-        this.updateCharList()
-        // window.addEventListener('scroll')
-    }
-
-    componentWillUnmount(){
-        // window.removeEventListener()
-    }
-
-    onCharListLoading = () => {
-        this.setState({
-            newItemLoading: true
-        })
-    }
-
-    onCharListLoaded = (newCharList) => {
+    const onCharListLoaded = (newCharList) => {
         let ended = false
         if (newCharList.length < 9) {
             ended = true
         }
         
-        
-        this.setState( ({offset, charList}) => ({
-                charList: [...charList, ...newCharList],
-                loading: false,
-                newItemLoading: false,
-                offset: offset + 9,
-                charEnded: ended,
-            }
-        ))
+        setCharList(charList => [...charList, ...newCharList])
+        setNewItemLoading(false)
+        setOffset(offset => offset + 9)
+        setCharEnded(ended)
     }
 
-    onError = () => {
-        this.setState({
-            loading: false, 
-            error: true
-        })
+    const updateCharList = (offset, initial) => {
+        initial ? setNewItemLoading(false) : setNewItemLoading(true);
+        getAllCharacters(offset)
+            .then(onCharListLoaded)
     }
 
-    updateCharList = (offset) => {
-        this.onCharListLoading()
-        this.marvelService
-            .getAllCharacters(offset)
-            .then(this.onCharListLoaded)
-            .catch(this.onError)
+    const refList = useRef([])
+
+    const focusOnItem = (id) => {
+        refList.current.forEach(item => item.classList.remove('char__item_selected'));
+        refList.current[id].classList.add('char__item_selected');
+        refList.current[id].focus();
     }
-
-    refList = []
-
-    refCreate = elem => {
-       this.refList.push(elem)       
-    }    
-
-    focusOnItem = (id) => {
-        this.refList.forEach(item => item.classList.remove('char__item_selected'));
-        this.refList[id].classList.add('char__item_selected');
-        this.refList[id].focus();
-    }
-    
-    renderItems = (charList) => {
+ 
+    const renderItems = (charList) => {
         const items = charList.map((item, i) => {           
             
             let styleImg = item.thumbnail.includes('image_not_available') ? 'unset' : 'cover'  
             styleImg = item.thumbnail.includes('4c002e0305708.gif') ? 'unset' : styleImg
 
             return (
-                <li 
-                    className="char__item"
-                    key={item.id}
-                    ref={this.refCreate}
-                    tabIndex="0"
-                    onClick={() => {
-                        this.props.onCharSelected(item.id);
-                        this.focusOnItem(i);
-                    }}
-                    onKeyPress={(e) => {
-                        if (e.key === ' ' || e.key === "Enter") {
-                            this.props.onCharSelected(item.id);
-                            this.focusOnItem(i);
-                    }}}>        
-                        <img src={item.thumbnail} alt={item.name} style={{objectFit: styleImg}}/>
-                        <div className="char__name">{item.name}</div>
-                </li>
+                <CSSTransition key={item.id} timeout={500} classNames="char__item">                    
+                    <li 
+                        className="char__item"
+                        key={item.id}
+                        ref={el => refList.current[i] = el}
+                        tabIndex="0"
+                        onClick={() => {
+                            props.onCharSelected(item.id);
+                            focusOnItem(i);
+                        }}
+                        onKeyPress={(e) => {
+                            if (e.key === ' ' || e.key === "Enter") {
+                                props.onCharSelected(item.id);
+                                focusOnItem(i);
+                        }}}>        
+                            <img src={item.thumbnail} alt={item.name} style={{objectFit: styleImg}}/>
+                            <div className="char__name">{item.name}</div>
+                    </li>
+                </CSSTransition>
             )
         })
     
         return (
-            <ul className="char__grid">
-                {items}
-            </ul>
+           <ul className='char__grid'>
+                <TransitionGroup component={null}>
+                    {items}            
+                </TransitionGroup>
+           </ul>
         )
     }
 
-    render() {
-        const {newItemLoading, charList, loading, error, offset, charEnded} = this.state
-        const items = this.renderItems(charList)
-        const errorMessage = error ? <ErrorMessage/> : null
-        const spinner = loading ? <Spinner/> : null
-        const content = !(loading || error) ? items : null
+    const items = renderItems(charList)
+    const errorMessage = error ? <ErrorMessage/> : null
+    const spinner = loading && !newItemLoading ? <Spinner/> : null
         
-        return (
-            <div className="char__list">               
-                    {errorMessage}
-                    {spinner}
-                    {content}                
-                <button 
-                    className="button button__main button__long"
-                    disabled={newItemLoading}
-                    style={{'display': charEnded ? 'none' : 'block'}}
-                    onClick={() => this.updateCharList(offset)}>
-                    <div className="inner">load more</div>
-                </button>
-            </div>
-        )
-    }
+    return (
+        <div className="char__list">               
+                {errorMessage}
+                {spinner}    
+                {items}           
+            <button 
+                className="button button__main button__long"
+                disabled={newItemLoading}
+                style={{'display': charEnded ? 'none' : 'block'}}
+                onClick={() => updateCharList(offset)}>
+                <div className="inner">load more</div>
+            </button>
+        </div>
+    )    
 }
 
-CharList.prosType = {
+CharList.propTypes = {
     onCharSelected: PropTypes.func.isRequired
 }
 
